Fix error on contacts with no contributions.
Regression caused by https://github.com/eileenmcnaughton/nz.co.fuzion.civitoken/pull/18 which incorrectly assumes the
hook cannot be called if the  contact has no contributions. Note it might still be broken  on soft credits
but since that was added in https://github.com/eileenmcnaughton/nz.co.fuzion.civitoken/pull/18 it's not a regression.

It might be  worth fixing for soft credits as long as a unit test is part of the fix